% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{const}
\alias{const}
\title{const}
\usage{
const(theta, tau)
}
\arguments{
\item{theta}{The parameter that controls how repulsive the DPP is}

\item{tau}{The parameter we fixed at a large number}
}
\value{
The computed normalizing constant
}
\description{
The function to compute the normalizing constant for Determinantal Point Process (DPP)
}
