% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{likeli_theta}
\alias{likeli_theta}
\title{likeli_theta}
\usage{
likeli_theta(theta, tau, Beta)
}
\arguments{
\item{theta}{The parameter that controls how repulsive the DPP is}

\item{tau}{The parameter we fixed at a large number}

\item{Beta}{The matrix that records the samples}
}
\value{
The likelihood
}
\description{
The function computes the likelihood of the hyperparameter
}
