% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_sigma_squre}
\alias{update_sigma_squre}
\title{update_sigma_squre}
\usage{
update_sigma_squre(Y, mu, a = 1, b = 1)
}
\arguments{
\item{Y}{The CAL observation matrix, with missing values}

\item{mu}{Current estimated mean matrix for CAL}

\item{a}{The hyper-parameter with default value being 1}

\item{b}{The hyper-parameter with default value being 1}
}
\value{
Updated noise variance
}
\description{
Update the noise variance
}
\seealso{
\link{update_RJ} for a complete example for all functions in this package.
}
