% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_theta_beta}
\alias{update_theta_beta}
\title{update_theta_beta}
\usage{
update_theta_beta(theta, tau, Beta, sig = 10)
}
\arguments{
\item{theta}{The DPP hyper-parameter for patient level}

\item{tau}{A fixed DPP hyper-parameter, which we suggest high value, say 10^5}

\item{Beta}{The linear coefficients for patient level covariates. Should be a matrix with S rows}

\item{sig}{The hyper-parameter with default value being 10}
}
\value{
updated DPP hyper-parameter for patient level
}
\description{
Update the DPP hyper-parameter for patient level
}
\seealso{
\link{update_RJ} for a complete example for all functions in this package.
}
