% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_theta_gamma}
\alias{update_theta_gamma}
\title{update_theta_gamma}
\usage{
update_theta_gamma(theta, tau, Gamma, S, Ds, sig = 10)
}
\arguments{
\item{theta}{The DPP hyper-parameter for patient level}

\item{tau}{A fixed DPP hyper-parameter, which we suggest high value, say 10^5}

\item{Gamma}{The linear coefficients for site level covariates. Should be a 3-dimensional array.}

\item{S}{The number of patient level clusters}

\item{Ds}{The number for site level clusters for each patient level cluster}

\item{sig}{The hyper-parameter with default value being 10}
}
\value{
updated DPP hyper-parameter for site level
}
\description{
Update the DPP hyper-parameter for site level
}
\seealso{
\link{update_RJ} for a complete example for all functions in this package.
}
